+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -456,6 +456,7 @@ sub build_isf_public_interface_contract {
             't/1287-isf-aggregate-drive-values.t',
             't/1288-isf-aggregate-drive-expression-values.t',
             't/1289-isf-aggregate-drive-call-values.t',
+            't/1290-isf-aggregate-drive-call-expression-values.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',